feat: Refactor email sending to use NodeMailerClient and remove email service
